アルソック作成開始

+++ Assets/Mansion/Scripts/Main/Entity/RoomData.cs
@@ -7,11 +7,12 @@ public class RoomData {
 
 	public int ItemCount{ get; set; }
 
-	public int ItemPrice{ get; set; }
+	public long ItemPrice{ get; set; }
 
-	public float GenerateSpeed{ get; set; }
+	public double GenerateSpeed{ get; set; }
 
 	public string ItemName{ get; set; }
 
 	public string ItemDescription{ get; set; }
+
 }
